Re-generate test fixtures with Ruby 2.1
This will make the tests pass on this Ruby version. They were previously
failing because Ruby < 2.2 doesn't have the `URI::RFC3986_Parser`
constant defined.

In the future we should perhaps add some extra error checking for now
and tell the user that they need to use the same Ruby version that was
used to create the cache files, but for now this will do.